+++ Utils/azure-toolkit-ide-libs/azure-toolkit-ide-common-lib/pom.xml
@@ -28,10 +28,6 @@
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.squareup.okhttp3</groupId>
-            <artifactId>okhttp</artifactId>
-        </dependency>
     </dependencies>
     <build>
         <plugins>
