+++ Utils/azure-toolkit-ide-libs/azure-toolkit-ide-springcloud-lib/pom.xml
@@ -32,5 +32,9 @@
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.squareup.okhttp3</groupId>
+            <artifactId>okhttp</artifactId>
+        </dependency>
     </dependencies>
 </project>
