+++ Utils/azure-toolkit-ide-libs/azure-toolkit-ide-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/ide/springcloud/portforwarder/WebSocketBuilder.java
@@ -1,9 +1,9 @@
 /*
  * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for license information.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.azure.toolkit.ide.common.util.remotedebugging;
+package com.microsoft.azure.toolkit.ide.springcloud.portforwarder;
 
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
